Verificación de autoría

+++ src/servicios/RutasPublicacion.py
@@ -1,29 +1,34 @@
-import io
 import json
 from http import HTTPStatus
 
-from flask import Blueprint, Response, request, send_file
+from flask import Blueprint, Response, request
 
 from src.negocio.Comentario import Comentario
 from src.negocio.Denuncia import Denuncia
-from src.negocio.Oferta import Oferta
 from src.negocio.Publicacion import Publicacion
 from src.negocio.Puntuacion import Puntuacion
 from src.servicios.Auth import Auth
-from src.transferencia_archivos.ServidorArchivos import ServidorArchivos
 
 rutas_publicacion = Blueprint("rutas_publicacion", __name__)
 
 
 @rutas_publicacion.route("/publicaciones/<idPublicacion>", methods=["DELETE"])
+@Auth.requires_token
 def eliminar_publicacion(idPublicacion):
-    status = Publicacion.eliminar_publicacion(idPublicacion)
+    publicacion = Publicacion()
+    publicacion.idPublicacion = idPublicacion
+    publicacion.obtener_autor_por_id()
+    token = request.headers.get("token")
+    if Auth.verificar_autor(publicacion.obtener_autor_por_id(), token):
+        status = Publicacion.eliminar_publicacion(idPublicacion)
+    else:
+        status = HTTPStatus.UNAUTHORIZED
+
     return Response(status=status)
 
 
-
-
 @rutas_publicacion.route("/publicaciones/<idPublicacion>/puntuaciones", methods=["POST"])
+@Auth.requires_token
 def puntuar_publicacion(idPublicacion):
     puntuacion_recibida = request.json
     valores_requeridos = {"idMiembro", "esPositiva"}
@@ -43,9 +48,8 @@ def puntuar_publicacion(idPublicacion):
 
 
 
-
-
 @rutas_publicacion.route("/publicaciones/<idPublicacion>/interaccion", methods=["GET"])
+@Auth.requires_token
 def obtener_interaccion(idPublicacion):
     respuesta = Response(status=HTTPStatus.OK)
     id_miembro_recibido = int(request.headers.get("idMiembro"))
@@ -58,7 +62,9 @@ def obtener_interaccion(idPublicacion):
         respuesta = Response(status=HTTPStatus.NOT_FOUND)
     return respuesta
 
+
 @rutas_publicacion.route("/publicaciones/<id_publicacion>/comentarios", methods=["POST"])
+@Auth.requires_token
 def registrar_comentario(id_publicacion):
     comentario_recibido = request.json
     valores_requeridos = {"idMiembro", "contenido"}
@@ -96,7 +102,9 @@ def obtener_comentarios(id_publicacion):
     return respuesta
 
 
+
 @rutas_publicacion.route("/publicaciones/<id_publicacion>/denuncias", methods=["POST"])
+@Auth.requires_token
 def registrar_denuncia(id_publicacion):
     denuncia_recibida = request.json
     valores_requeridos = {"idDenunciante", "comentario", "motivo"}
